+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 0.X.X
 
+- Function signature change:
+  - from `watch(in:watch:) throws -> EventLoopFuture<Void>`
+  - to `watch(in:using:) throws -> HTTPClient.Task<Void>`
+- Function signature change:
+  -  from`follow(in:name:container:watch:) throws -> HTTPClient.Task<Void>`
+  -  to `follow(in:name:container:using:) throws -> HTTPClient.Task<Void>`
+
 ## 0.1.0
 
 Initial release